agent-app/pyproject.toml
[project]
name = "mobile-agent"
version = "0.1.0"
description = "基于 langchain.agents + langgraph-agent-kit 的移动端自动化 Agent"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Agent 核心（langchain 封装了 langgraph）
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.4",

    # 流式编排 SDK
    "langgraph-agent-kit",

    # MCP 客户端
    "langchain-mcp-adapters>=0.1",
    "mcp>=1.0",

    # HTTP API
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "sse-starlette>=2.0",

    # 持久化
    "aiosqlite>=0.20",
    "langgraph-checkpoint-sqlite>=2.0",

    # 工具
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[tool.uv.sources]
langgraph-agent-kit = { path = "/Users/wang/code/my/embedease-sdk/backend/langgraph-agent-kit" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mobile_agent"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[project.scripts]
mobile-agent = "mobile_agent.cli.interactive:main"
